# PIL-470 Activity feed performance improvements