… package.json
… src/components/ActivityFeed/__tests__/ActivityFeed.test.js
… src/components/ActivityFeed/__tests__/__snapshots__/ActivityFeed.test.js.snap
… src/screens/Home/Home.js
… src/selectors/selectors.js
… yarn.lock